add view for seeing instructions and address etc for a given instance (also pylint is being weird)
