+++ src/pttbackend/api.py
@@ -10,6 +10,7 @@ from arkia11napi.middleware import DBWrapper
 from .config import STATIC_PATH, LOG_LEVEL
 from .views.instances import INSTANCE_ROUTER
 from .views.callbacks import CALLBACKS_ROUTER
+from .views.instructions import INSTRUCTIONS_ROUTER
 
 from . import models
 
@@ -17,6 +18,7 @@ from . import models
 LOGGER = logging.getLogger(__name__)
 APP = FastAPI(docs_url="/api/docs", openapi_url="/api/openapi.json")
 APP.mount("/static", StaticFiles(directory=str(STATIC_PATH)), name="static")
+APP.include_router(INSTRUCTIONS_ROUTER)
 APP.include_router(CALLBACKS_ROUTER)
 APP.include_router(INSTANCE_ROUTER)
 WRAPPER = DBWrapper(gino=models.db)
