+++ src/pttbackend/templates/instructions.html
@@ -9,9 +9,9 @@
   <h1>PTT Ohjeet</h1>
   <h2>Dokumentit</h2>
     <ul>
-      <li><a href="{{ instructions_pdf }}}">Käyttöohjeet</a></li>
-      <li><a href="{{ taisteluajatus_pdf }}}">Taisteluajatus</a></li>
-      <li><a href="{{ templates_zip }}}">Dokumenttipohjat</a></li>
+      <li><a href="{{ instructions_pdf }}">Käyttöohjeet</a></li>
+      <li><a href="{{ taisteluajatus_pdf }}">Taisteluajatus</a></li>
+      <li><a href="{{ templates_zip }}">Dokumenttipohjat</a></li>
     </ul>
   <h2>Viestiperusteet</h2>
     <table>
