+++ src/pttbackend/views/instances.py
@@ -4,8 +4,7 @@ import logging
 import uuid
 
 import pendulum
-from fastapi import APIRouter, Depends, Request, HTTPException, Response
-from fastapi.responses import HTMLResponse
+from fastapi import APIRouter, Depends, Request, HTTPException
 from fastapi.templating import Jinja2Templates
 from starlette import status
 from arkia11napi.helpers import get_or_404
@@ -16,7 +15,6 @@ from ..config import TEMPLATES_PATH
 from ..schemas.instance import DBInstance, InstanceCreate, InstancePager
 from ..models import PTTInstance
 from ..pipelineclient import PipeLineClient
-from .. import config
 
 
 LOGGER = logging.getLogger(__name__)
@@ -110,24 +108,3 @@ async def delete_instance(request: Request, pkstr: str) -> None:
         LOGGER.exception("Could not trigger pipeline {}".format(exc))
         raise
     await instance.update(deleted=pendulum.now("UTC")).apply()
-
-
-@INSTANCE_ROUTER.get("/api/v1/ptt/instances/{pkstr}/instructions", tags=["ptt-instances"], response_class=HTMLResponse)
-async def get_instructions(request: Request, pkstr: str) -> Response:
-    """Show instructions HTML with server address etc"""
-    instance = await get_or_404(PTTInstance, pkstr)
-    if not check_acl(request.state.jwt, "fi.pvarki.pttbackend.instance:read", auto_error=False):
-        if instance.ownerid != request.state.jwt["userid"]:
-            raise HTTPException(status_code=403, detail="Required privilege not granted.")
-    return TEMPLATES.TemplateResponse(
-        "instructions.html",
-        {
-            "request": request,
-            "instructions_pdf": config.INSTRUCTIONS_URL,
-            "taisteluajatus_pdf": config.TAKORTTI_URL,
-            "templates_zip": config.DOCTEMPLATE_URL,
-            "dns_name": instance.tfoutputs["dns_name"],
-            "mumble_server_password": instance.tfoutputs["mumble_server_password"],
-            "mumble_superuser_password": instance.tfoutputs["mumble_superuser_password"],
-        },
-    )
